+++ app/src/debug/res/values/google_maps_api.xml
@@ -1,24 +1,3 @@
 <resources>
-    <!--
-    TODO: Before you run your application, you need a Google Maps API key.
-
-    To get one, follow this link, follow the directions and press "Create" at the end:
-
-    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=41:13:45:26:F8:A5:27:0D:E1:58:4E:78:6F:B1:E5:3D:CF:3F:0C:C2%3Bcom.kashanok.cleancode.presentation.screen.vehicle
-
-    You can also add your credentials to an existing key, using these values:
-
-    Package name:
-    41:13:45:26:F8:A5:27:0D:E1:58:4E:78:6F:B1:E5:3D:CF:3F:0C:C2
-
-    SHA-1 certificate fingerprint:
-    41:13:45:26:F8:A5:27:0D:E1:58:4E:78:6F:B1:E5:3D:CF:3F:0C:C2
-
-    Alternatively, follow the directions here:
-    https://developers.google.com/maps/documentation/android/start#get-key
-
-    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
-    string in this file.
-    -->
     <string name="google_maps_key" translatable="false" templateMergeStrategy="preserve">AIzaSyBz3GRsZQq4FsZRcb5NdWLUYY2AKkMV_lI</string>
 </resources>
