+++ app/src/main/java/com/kashanok/cleancode/app/UseCaseProvider.kt
@@ -2,6 +2,7 @@ package com.kashanok.cleancode.app
 
 import com.kashanok.cleancode.BuildConfig
 import com.kashanok.data.repository.VehicleRepositoryDefault
+import com.kashanok.data.rest.VehicleService
 import com.kashanok.domain.usecase.vehicle.VehicleGetUseCase
 import com.kashanok.domain.usecase.vehicle.VehicleGetUseCaseDefault
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -14,8 +15,8 @@ object UseCaseProvider {
             getWorkScheduler(),
             getPostScheduler(),
             VehicleRepositoryDefault(
-                VehicleApplication.instance.applicationContext,
-                BuildConfig.API_ENDPOINT
+                BuildConfig.API_ENDPOINT,
+                VehicleService(VehicleApplication.instance.applicationContext)
             )
         )
     }
