+++ app/src/main/java/com/kashanok/cleancode/presentation/base/BaseMvvmFragment.kt
@@ -30,6 +30,5 @@ abstract class BaseMvvmFragment<VM: BaseViewModel> : Fragment() {
         initView()
     }
 
-
     abstract fun initView()
 }
