+++ app/src/main/res/layout/fragment_vehicle_list.xml
@@ -3,7 +3,7 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="match_parent" xmlns:tools="http://schemas.android.com/tools"
         android:id="@+id/mapList">
 
     <FrameLayout
@@ -12,6 +12,14 @@
             android:layout_height="match_parent"
             android:background="@color/colorWhite"
             app:layout_behavior="@string/bottom_sheet_behavior">
+
+        <ProgressBar
+                android:id="@+id/progressBar"
+                android:visibility="invisible"
+                tools:visibility="visible"
+                android:layout_gravity="center_horizontal"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content" />
 
         <android.support.v7.widget.RecyclerView
                 android:id="@+id/rvVehicles"
