+++ app/src/release/res/values/google_maps_api.xml
@@ -1,20 +1,3 @@
 <resources>
-    <!--
-    TODO: Before you release your application, you need a Google Maps API key.
-
-    To do this, you can either add your release key credentials to your existing
-    key, or create a new key.
-
-    Note that this file specifies the API key for the release build target.
-    If you have previously set up a key for the debug target with the debug signing certificate,
-    you will also need to set up a key for your release certificate.
-
-    Follow the directions here:
-
-    https://developers.google.com/maps/documentation/android/signup
-
-    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
-    string in this file.
-    -->
     <string name="google_maps_key" translatable="false" templateMergeStrategy="preserve">AIzaSyBz3GRsZQq4FsZRcb5NdWLUYY2AKkMV_lI</string>
 </resources>
